transporter serializer, and remove transporter from Manifest model to separate model
