+++ apps/api/serializers/__init__.py
@@ -1,4 +1,4 @@
 from .addresses import MailAddressField, SiteAddressField
-from .handlers import HandlerSerializer
-from .manifests import ManifestSerializer, WasteLineSerializer
-from .transporters import TransporterSerializer
+from .handler import HandlerSerializer
+from .manifest import (ManifestSerializer, TransporterSerializer,
+                       WasteLineSerializer)
