+++ apps/api/serializers/handler.py
@@ -68,9 +68,7 @@ class HandlerSerializer(serializers.ModelSerializer):
             'mailingAddress',
             'contact',
             'emergencyPhone',
-            # 'paperSignatureInfo',
             'electronicSignatureInfo',
-            # 'order',
             'registered',
             'limitedEsign',
             'canEsign',
