--- apps/api/serializers/transporters.py
@@ -1,10 +1,0 @@
-from apps.trak.models import Transporter
-
-from . import HandlerSerializer
-
-
-class TransporterSerializer(HandlerSerializer):
-    class Meta:
-        model = Transporter
-        fields = '__all__'
-        depth = 3
