+++ apps/trak/models/transporter.py
@@ -1,9 +1,14 @@
 from django.db import models
 
-from . import Handler
+from . import Handler, Manifest
 
 
 class Transporter(Handler):
+    manifest = models.ForeignKey(
+        Manifest,
+        related_name='transporters',
+        on_delete=models.CASCADE,
+    )
     order = models.IntegerField()
 
     def __str__(self):
